+++ docs/CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- extend flaw-task linking to primarily use the CVE ID
+
+### Fixed
+- fix OSIDB-Bugzilla mid-air collision issues (OSIDB-3083)
+
 ## [4.1.1] - 2024-06-28
 ### Added
 - prefetch Alerts related models for each API endpoint (OSIDB-3053)